add: layout style para o dashboard
