+++ gotravel-frontend/package.json
@@ -11,6 +11,7 @@
     "axios": "^0.20.0",
     "date-fns": "^2.16.1",
     "react": "^16.13.1",
+    "react-cool-onclickoutside": "^1.5.7",
     "react-dom": "^16.13.1",
     "react-google-maps": "^9.4.5",
     "react-hook-form": "^6.9.0",
