+++ gotravel-frontend/src/components/Organisms/ContainerMaps/style.css
@@ -1,0 +1,4 @@
+.map-container {
+  width: 100vw;
+  height: 400px;
+}
