+++ gotravel-frontend/src/components/Organisms/Login/Login.js
@@ -1,5 +1,5 @@
 import React, { useContext, useState } from 'react';
-import { BrowserRouter, Switch, Route } from 'react-router-dom';
+import { BrowserRouter, Switch, Route, Redirect } from 'react-router-dom';
 
 import FormLogin from '../../Molecules/FormLogin/FormLogin';
 import FormRegister from '../../Molecules/FormRegister/FormRegister';
@@ -9,9 +9,13 @@ import StateContext from '../../../StateContext';
 import DispatchContext from '../../../DispatchContext';
 import Layout from '../../Layout';
 
-const Login = () => {
+const Login = (props) => {
   const appState = useContext(StateContext);
   const appDispatch = useContext(DispatchContext);
+
+  if (appState.logged) {
+    return <Redirect to="/dashboard" />;
+  }
 
   return <FirstContainer>{appState.formType ? <FormLogin /> : <FormRegister />}</FirstContainer>;
 };
