+++ gotravel-frontend/src/pages/dashboard/dashboard.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import Layout from '../../components/Layout';
+import ContainerMaps from '../../components/Organisms/ContainerMaps/ContainerMaps';
+
+const Dashboard = () => {
+  return (
+    <Layout>
+      <ContainerMaps />
+    </Layout>
+  );
+};
+
+export default Dashboard;
