--- gotravel-frontend/src/pages/dashboard/dashboard.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import ContainerMaps from '../../components/Organisms/ContainerMaps/ContainerMaps';
-
-const Dashboard = () => {
-
-    return (
-        <ContainerMaps/> 
-    );
-}
-
-export default Dashboard;
